Move script_plugins to a single file
